requirements.txt
fastapi
uvicorn
pandas
pyyaml
python-multipart